+++ DAL/Interfaces/IEmbedding.cs
@@ -2,5 +2,5 @@ namespace Domain.Interfaces;
 
 public interface IEmbedding
 {
-    float[] Embedding { get; }    
+    byte[] Embedding { get; }    
 }
